pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-mol"
version = "0.1.0"
description = "SMILES + temperature -> aqueous diffusion coefficient"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "scikit-learn>=1.3",
    "torch>=2.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "rdkit>=2023.3.1",
    "torch-geometric>=2.4",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
